Add keyboard and audio description support
Keyboard to match colors of correct, incorrect, misplaced, and unused letters. You can also click on the keys to enter them into guesses in addition to typing.

Option to turn on audio descriptions. Turn this on by right clicking. When this option is on, descriptions of the keys will be read - the letter of the key and whether it's correct, misplaced, incorrect, or unused.

To do: Change screen reader descriptions to audio descriptions, since this isn't really a screen reader. Also add support for a clickable enter button and reading the results out loud.